refactor(blocks-construct): chat feature revamp
