+++ src/features/chat/components/chat-users/chat-users.tsx
@@ -1,19 +1,6 @@
 import { useState, useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  BellOff,
-  EllipsisVertical,
-  Info,
-  Phone,
-  Reply,
-  Smile,
-  Trash,
-  Users,
-  Video,
-  Bell,
-  FileText,
-  Download,
-} from 'lucide-react';
+import { BellOff, EllipsisVertical, Info, Phone, Trash, Users, Video, Bell } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { Separator } from '@/components/ui/separator';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
@@ -28,152 +15,8 @@ import { ForwardMessage } from '../modals/forward-message/forward-message';
 import { ChatProfile } from '../chat-profile/chat-profile';
 import { ChatInput } from '../chat-input/chat-input';
 import { ChatContact, Message } from '../../types/chat.types';
-
-const MessageMenu = ({
-  msg,
-  openDropdownId,
-  setOpenDropdownId,
-  handleOpenForwardModal,
-  handleDeleteMessage,
-  t,
-}: {
-  msg: Message;
-  openDropdownId: string | null;
-  setOpenDropdownId: (id: string | null) => void;
-  handleOpenForwardModal: (msg: Message) => void;
-  handleDeleteMessage: (id: string) => void;
-  t: (key: string) => string;
-}) => (
-  <DropdownMenu
-    open={openDropdownId === msg.id}
-    onOpenChange={(open) => setOpenDropdownId(open ? msg.id : null)}
-  >
-    <DropdownMenuTrigger asChild>
-      <Button variant="ghost" size="icon" className="h-6 w-6 p-0.5 rounded-full">
-        <EllipsisVertical className="w-4 h-4" />
-      </Button>
-    </DropdownMenuTrigger>
-    <DropdownMenuContent className="w-40" align="end">
-      <DropdownMenuItem
-        onClick={() => handleOpenForwardModal(msg)}
-        data-testid={`msg-${msg.id}-forward-btn`}
-      >
-        <Reply className="w-4 h-4 mr-2" />
-        {t('FORWARD')}
-      </DropdownMenuItem>
-      <DropdownMenuItem
-        onClick={() => handleDeleteMessage(msg.id)}
-        data-testid={`msg-${msg.id}-delete-btn`}
-      >
-        <Trash className="w-4 h-4 mr-2" />
-        {t('DELETE')}
-      </DropdownMenuItem>
-    </DropdownMenuContent>
-  </DropdownMenu>
-);
-
-const MessageContent = ({
-  msg,
-  formatFileSize,
-  handleDownload,
-}: {
-  msg: Message;
-  formatFileSize: (size: number) => string;
-  handleDownload: (file: { url: string; name: string }) => void;
-}) => (
-  <>
-    <p className="text-sm text-high-emphasis">{msg.content}</p>
-    {msg.attachment && (
-      <div className="mt-2">
-        {msg.attachment.type.startsWith('image/') ? (
-          <img
-            src={msg.attachment.url}
-            alt={msg.attachment.name}
-            className="max-w-full max-h-64 rounded-lg border border-border"
-          />
-        ) : (
-          <div className="flex items-center gap-4 w-full">
-            <div className="flex items-center justify-center w-10 h-10 bg-white rounded-[4px]">
-              <FileText className="w-6 h-6 text-secondary" />
-            </div>
-            <div className="flex flex-col">
-              <div className="text-sm text-medium-emphasis truncate">{msg.attachment.name}</div>
-              <div className="text-xs text-low-emphasis">{formatFileSize(msg.attachment.size)}</div>
-            </div>
-            {msg.attachment &&
-              (() => {
-                const { url, name } = msg.attachment;
-                return (
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    className="rounded-full"
-                    onClick={() => handleDownload({ url, name })}
-                    aria-label="Download"
-                  >
-                    <Download className="w-5 h-5 text-medium-emphasis" />
-                  </Button>
-                );
-              })()}
-          </div>
-        )}
-      </div>
-    )}
-  </>
-);
-
-const MessageActions = ({
-  isMe,
-  msg,
-  openDropdownId,
-  setOpenDropdownId,
-  handleOpenForwardModal,
-  handleDeleteMessage,
-  t,
-}: {
-  isMe: boolean;
-  msg: Message;
-  openDropdownId: string | null;
-  setOpenDropdownId: (id: string | null) => void;
-  handleOpenForwardModal: (msg: Message) => void;
-  handleDeleteMessage: (id: string) => void;
-  t: (key: string) => string;
-}) => {
-  const actions = [
-    <Button key="smile" variant="ghost" size="icon" className="h-6 w-6 p-0.5 rounded-full">
-      <Smile className="w-4 h-4" />
-    </Button>,
-    <Button
-      key="reply"
-      variant="ghost"
-      size="icon"
-      className="h-6 w-6 p-0.5 rounded-full"
-      onClick={() => handleOpenForwardModal(msg)}
-    >
-      <Reply className="w-4 h-4" />
-    </Button>,
-    <MessageMenu
-      key="menu"
-      msg={msg}
-      openDropdownId={openDropdownId}
-      setOpenDropdownId={setOpenDropdownId}
-      handleOpenForwardModal={handleOpenForwardModal}
-      handleDeleteMessage={handleDeleteMessage}
-      t={t}
-    />,
-  ];
-
-  return (
-    <div
-      className={cn(
-        'flex items-center gap-1 transition-opacity duration-200',
-        openDropdownId === msg.id ? 'opacity-100' : 'opacity-0 group-hover:opacity-100'
-      )}
-    >
-      {isMe ? actions.slice().reverse() : actions}
-    </div>
-  );
-};
+import { MessageActions } from '../message-actions/message-actions';
+import { MessageContent } from '../message-content/message-content';
 
 interface ChatUsersProps {
   contact: ChatContact;
