+++ src/features/chat/services/chat.services.ts
@@ -392,3 +392,26 @@ export const mockChatContacts: ChatContact[] = [
     messages: dhanvirMessages,
   }),
 ];
+
+export const getAttachmentBackgroundColor = (type: string): string => {
+  switch (type) {
+    case 'pdf':
+      return 'bg-blue-50';
+    case 'image':
+      return 'bg-red-50';
+    case 'audio':
+      return 'bg-purple-50';
+    case 'video':
+      return 'bg-green-50';
+    default:
+      return 'bg-surface';
+  }
+};
+
+export const attachments = [
+  { id: '1', name: 'acceptance criteria final.pdf', size: '600.00 KB', type: 'pdf' },
+  { id: '2', name: 'Sunset_View_Image.jpg', size: '600.00 KB', type: 'image' },
+  { id: '3', name: 'acceptance criteria preview vers...', size: '600.00 KB', type: 'pdf' },
+  { id: '4', name: 'discussion.mp3', size: '600.00 KB', type: 'audio' },
+  { id: '5', name: 'meeting_notes.mp4', size: '500.00 MB', type: 'video' },
+];
